adding seed data and test ActiveRecord Associations

+++ db/migrate/20200706145652_create_runners.rb
@@ -3,7 +3,7 @@ class CreateRunners < ActiveRecord::Migration[6.0]
     create_table :runners do |t|
       t.string :name
       t.integer :age
-      t.float :height
+      t.integer :height
       t.integer :weight
       t.float :shoe_size
